requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
flask-cors==4.0.0